Simplify helper wrapper
This improves performance slightly and gives the wrapper function a more
descriptive name to help with debugging/profiling.
