+++ lib/handlebars/helpers.js
@@ -25,3 +25,32 @@ export function moveHelperToHooks(instance, helperName, keepHelper) {
     }
   }
 }
+
+export function mergeHelpers(env, options, container) {
+  const mergedHelpers = {};
+  addHelpers(mergedHelpers, env.helpers, container);
+  addHelpers(mergedHelpers, options.helpers, container);
+  return mergedHelpers;
+}
+
+function addHelpers(mergedHelpers, helpers, container) {
+  if (!helpers) return;
+  Object.keys(helpers).forEach((helperName) => {
+    const helper = helpers[helperName];
+    mergedHelpers[helperName] = passLookupPropertyOption(helper, container);
+  });
+}
+
+function passLookupPropertyOption(helper, container) {
+  if (typeof helper !== 'function') {
+    // This should not happen, but apparently it does in https://github.com/wycats/handlebars.js/issues/1639
+    // We try to make the wrapper least-invasive by not wrapping it, if the helper is not a function.
+    return helper;
+  }
+  const lookupProperty = container.lookupProperty;
+  return function invokeHelper(/* dynamic arguments */) {
+    const options = arguments[arguments.length - 1];
+    options.lookupProperty = lookupProperty;
+    return helper.apply(this, arguments);
+  };
+}
