+++ lib/handlebars/runtime.js
@@ -6,8 +6,7 @@ import {
   LAST_COMPATIBLE_COMPILER_REVISION,
   REVISION_CHANGES,
 } from './base.js';
-import { moveHelperToHooks } from './helpers.js';
-import { wrapHelper } from './internal/wrapHelper.js';
+import { mergeHelpers, moveHelperToHooks } from './helpers.js';
 import {
   createProtoAccessControl,
   resultIsAllowed,
@@ -264,10 +263,7 @@ export function template(templateSpec, env) {
 
   function _setup(options) {
     if (!options.partial) {
-      let mergedHelpers = {};
-      addHelpers(mergedHelpers, env.helpers, container);
-      addHelpers(mergedHelpers, options.helpers, container);
-      container.helpers = mergedHelpers;
+      container.helpers = mergeHelpers(env, options, container);
 
       if (templateSpec.usePartial) {
         // Use mergeIfNeeded here to prevent compiling global partials multiple times
@@ -425,20 +421,4 @@ function executeDecorators(fn, prog, container, depths, data, blockParams) {
     Utils.extend(prog, props);
   }
   return prog;
-}
-
-function addHelpers(mergedHelpers, helpers, container) {
-  if (!helpers) return;
-  Object.keys(helpers).forEach((helperName) => {
-    let helper = helpers[helperName];
-    mergedHelpers[helperName] = passLookupPropertyOption(helper, container);
-  });
-}
-
-function passLookupPropertyOption(helper, container) {
-  const lookupProperty = container.lookupProperty;
-  return wrapHelper(helper, (options) => {
-    options.lookupProperty = lookupProperty;
-    return options;
-  });
 }
